test: suppress a few warnings (memory)

+++ test/entt/core/memory.cpp
@@ -60,8 +60,8 @@ TEST(NextPowerOfTwo, Functionalities) {
     ASSERT_EQ(entt::next_power_of_two(17u), 32u);
     ASSERT_EQ(entt::next_power_of_two(32u), 32u);
     ASSERT_EQ(entt::next_power_of_two(33u), 64u);
-    ASSERT_EQ(entt::next_power_of_two(std::pow(2, 16)), std::pow(2, 16));
-    ASSERT_EQ(entt::next_power_of_two(std::pow(2, 16) + 1u), std::pow(2, 17));
+    ASSERT_EQ(entt::next_power_of_two(static_cast<std::size_t>(std::pow(2, 16))), static_cast<std::size_t>(std::pow(2, 16)));
+    ASSERT_EQ(entt::next_power_of_two(static_cast<std::size_t>(std::pow(2, 16) + 1u)), static_cast<std::size_t>(std::pow(2, 17)));
 }
 
 ENTT_DEBUG_TEST(NextPowerOfTwoDeathTest, Functionalities) {
